feat: mainPage와 transactionList 파일 분리
